Improve reporting when cassette does not contain response
It may be tricky to determine why a recorded response is not used as no
information about how the actual request compares to the recorded once.

This change changes the behaviour of the playback method to throw an 
exception is a match is not found. The message contains information 
about the involved requests.

If PHP-VCR is also configured not to allow recording of new requests 
then the new exception is chained to the existing one.

Additional information related to request matching is added to the new
root exception.